benchmarks/requirements.txt
psycopg2-binary==2.9.*
prometheus_client==0.20.*
sqlparse==0.5.*
